+++ lib/karafka/web/contracts/config.rb
@@ -52,6 +52,7 @@ module Karafka
             required(:secret) { |val| val.is_a?(String) && val.length >= 64 }
           end
 
+          required(:show_internal_topics) { |val| [true, false].include?(val) }
           required(:cache) { |val| !val.nil? }
           required(:per_page) { |val| val.is_a?(Integer) && val >= 1 && val <= 100 }
           required(:visibility_filter) { |val| !val.nil? }