+++ lib/karafka/web/tracking/consumers/contracts/partition.rb
@@ -15,7 +15,9 @@ module Karafka
             required(:lag_stored) { |val| val.is_a?(Integer) }
             required(:lag_stored_d) { |val| val.is_a?(Integer) }
             required(:committed_offset) { |val| val.is_a?(Integer) }
+            required(:committed_offset_fd) { |val| val.is_a?(Integer) && val >= 0 }
             required(:stored_offset) { |val| val.is_a?(Integer) }
+            required(:stored_offset_fd) { |val| val.is_a?(Integer) && val >= 0 }
             required(:fetch_state) { |val| val.is_a?(String) && !val.empty? }
             required(:poll_state) { |val| val.is_a?(String) && !val.empty? }
             required(:hi_offset) { |val| val.is_a?(Integer) }