+++ spec/lib/karafka/web/contracts/config_spec.rb
@@ -37,6 +37,7 @@ RSpec.describe_current do
           key: 'some_key',
           secret: 'a' * 64
         },
+        show_internal_topics: true,
         cache: Object.new,
         per_page: 50,
         visibility_filter: Object.new
@@ -165,6 +166,18 @@ RSpec.describe_current do
 
     context 'when visibility_filter is nil' do
       before { params[:ui][:visibility_filter] = nil }
+
+      it { expect(contract.call(params)).not_to be_success }
+    end
+
+    context 'when show_internal_topics is nil' do
+      before { params[:ui][:show_internal_topics] = nil }
+
+      it { expect(contract.call(params)).not_to be_success }
+    end
+
+    context 'when show_internal_topics is not boolean' do
+      before { params[:ui][:show_internal_topics] = '1' }
 
       it { expect(contract.call(params)).not_to be_success }
     end