+++ spec/lib/karafka/web/tracking/consumers/contracts/partition_spec.rb
@@ -11,7 +11,9 @@ RSpec.describe_current do
       lag: 2,
       lag_d: -1,
       committed_offset: 0,
+      committed_offset_fd: 0,
       stored_offset: 0,
+      stored_offset_fd: 0,
       fetch_state: 'active',
       poll_state: 'active',
       hi_offset: 1,
@@ -63,7 +65,9 @@ RSpec.describe_current do
     lag
     lag_d
     committed_offset
+    committed_offset_fd
     stored_offset
+    stored_offset_fd
     hi_offset
     lo_offset
     eof_offset
@@ -84,9 +88,15 @@ RSpec.describe_current do
     end
   end
 
-  context 'when ls_offset_fd is less than 0' do
-    before { config[:ls_offset_fd] = -1 }
+  %i[
+    committed_offset_fd
+    stored_offset_fd
+    ls_offset_fd
+  ].each do |fd|
+    context "when #{fd} is less than 0" do
+      before { config[fd] = -1 }
 
-    it { expect(contract.call(config)).not_to be_success }
+      it { expect(contract.call(config)).not_to be_success }
+    end
   end
 end